frontend: refactor NPO wizard draft flow
- remove unused NPO auth fields from wizard state persistence
- collapse the NPO flow to the current 4-step flow and drop Step 5
- make the replayed first-project payload respect the selected project type
