+++ frontend/src/pages/RegisterPage/Steps/Step4.tsx
@@ -1,12 +1,11 @@
-import { useEffect } from "react";
 import type { Dispatch, SetStateAction } from "react";
 import { Input } from "../../../components/general/Input";
-import { AddressIcon, StateIcon, PhoneIcon } from "../../../components/icons";
-import { useZipCode } from "../../../hooks/useZipCode";
-import { formatZipCode } from "../../../utils/formatZipCode";
-import { InfoBox } from "../../../components/general/InfoBox";
-import type { FieldErrors, WizardFormData } from "../../../types/wizard.types";
-import { getApiErrorMessage } from "../../../utils/logger";
+import { TextArea } from "../../../components/general/TextArea";
+import type {
+  FieldErrors,
+  ProjectOdsOption,
+  WizardFormData,
+} from "../../../types/wizard.types";
 
 type Step4Props = {
   formData: WizardFormData;
@@ -14,169 +13,214 @@ type Step4Props = {
   errors: FieldErrors;
 };
 
+const ODS_OPTIONS: Array<{
+  value: ProjectOdsOption;
+  label: string;
+  description: string;
+}> = [
+  {
+    value: "1",
+    label: "ODS 1 - Erradicação da Pobreza",
+    description: "Projetos de redução de desigualdades e vulnerabilidade.",
+  },
+  {
+    value: "2",
+    label: "ODS 2 - Fome Zero",
+    description:
+      "Iniciativas de segurança alimentar e agricultura sustentável.",
+  },
+  {
+    value: "3",
+    label: "ODS 3 - Saúde e Bem-Estar",
+    description: "Ações voltadas à saúde, prevenção e qualidade de vida.",
+  },
+];
+
+const PROJECT_TYPE_OPTIONS: Array<{
+  value: Exclude<WizardFormData["tipoProjeto"], "">;
+  label: string;
+  description: string;
+}> = [
+  {
+    value: "social",
+    label: "Social",
+    description: "Não exige meta de captação.",
+  },
+  {
+    value: "governamental",
+    label: "Governamental",
+    description: "Exige meta de captação.",
+  },
+];
+
 export function Step4({ formData, setFormData, errors }: Step4Props) {
-  const {
-    data: zipCodeData,
-    isFetching: loadingZipCode,
-    error: zipCodeQueryError,
-  } = useZipCode(formData.zipCode);
-
-  useEffect(() => {
-    if (zipCodeData) {
-      setFormData((prev) => ({
-        ...prev,
-        street: zipCodeData.street || prev.street,
-        streetNumber: (zipCodeData.complement ?? "").slice(0, 20) || prev.streetNumber,
-        city: zipCodeData.city || prev.city,
-        state: zipCodeData.state || prev.state,
-        stateCode: zipCodeData.stateCode || prev.stateCode,
-      }));
-    }
-  }, [zipCodeData, setFormData]);
-
-  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const { id, value } = e.target;
-    const formatted = id === "zipCode" ? formatZipCode(value) : value;
-
-    if (id === "zipCode") {
-      setFormData((prev) => ({
-        ...prev,
-        zipCode: formatted,
-        street: "",
-        city: "",
-        state: "",
-        stateCode: "",
-      }));
-      return;
-    }
-
-    setFormData((prev) => ({ ...prev, [id]: formatted }));
+  function toggleOds(value: ProjectOdsOption) {
+    setFormData((prev) => ({
+      ...prev,
+      odsProjeto: prev.odsProjeto.includes(value)
+        ? prev.odsProjeto.filter((item) => item !== value)
+        : [...prev.odsProjeto, value],
+    }));
   }
 
   return (
     <div className="flex flex-col gap-6">
       <div>
-        <h2 className="text-vinculo-dark font-semibold text-lg">Endereço</h2>
+        <h2 className="text-vinculo-dark font-semibold text-lg">
+          Primeiro projeto
+        </h2>
         <p className="text-sm text-slate-500 mt-1">
-          Preencha os dados de localização da sua ONG.
+          Informe o projeto inicial que será cadastrado junto com a ONG.
         </p>
       </div>
 
-      <div className="flex flex-col gap-4">
-        <div className="flex flex-col gap-1">
-          <Input
-            label="CEP"
-            id="zipCode"
-            placeholder="00000-000"
-            maxLength={9}
-            value={formData.zipCode}
-            onChange={handleChange}
-            error={errors.zipCode}
-            icon={<AddressIcon />}
-            iconPosition="left"
-            isRequired
-          />
-          {loadingZipCode && <span className="text-sm text-slate-400">Consultando CEP...</span>}
-          {zipCodeQueryError && (
-            <span className="text-sm text-error">
-              {getApiErrorMessage(zipCodeQueryError, "CEP não encontrado. Verifique e tente novamente.")}
-            </span>
-          )}
-        </div>
+      <Input
+        id="nomeProjeto"
+        label="Nome do projeto"
+        isRequired
+        placeholder="Ex: Educação para Todos"
+        maxLength={255}
+        value={formData.nomeProjeto}
+        onChange={(e) =>
+          setFormData((prev) => ({
+            ...prev,
+            nomeProjeto: e.target.value,
+          }))
+        }
+        error={errors.nomeProjeto}
+      />
 
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-          <div className="md:col-span-2">
-            <Input
-              label="Logradouro"
-              id="street"
-              placeholder="Rua, Avenida..."
-              maxLength={100}
-              value={formData.street}
-              onChange={handleChange}
-              icon={<AddressIcon />}
-              iconPosition="left"
-              disabled={!!zipCodeData}
-            />
-          </div>
-          <Input
-            label="Número"
-            id="streetNumber"
-            placeholder="Ex: 123"
-            maxLength={20}
-            value={formData.streetNumber}
-            onChange={handleChange}
-            error={errors.streetNumber}
-            icon={<AddressIcon />}
-            iconPosition="left"
-            isRequired
-          />
-        </div>
+      <TextArea
+        id="descricaoProjeto"
+        label="Descrição do projeto"
+        isRequired
+        placeholder="Descreva o objetivo, impacto e público beneficiado."
+        maxLength={1000}
+        value={formData.descricaoProjeto}
+        onChange={(e) =>
+          setFormData((prev) => ({
+            ...prev,
+            descricaoProjeto: e.target.value,
+          }))
+        }
+      />
+      {errors.descricaoProjeto && (
+        <p className="text-sm text-error -mt-4" role="alert">
+          {errors.descricaoProjeto}
+        </p>
+      )}
 
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          <Input
-            label="Complemento"
-            id="complement"
-            placeholder="Apto, Sala..."
-            maxLength={100}
-            value={formData.complement}
-            onChange={handleChange}
-            icon={<AddressIcon />}
-            iconPosition="left"
-          />
-          <Input
-            label="Cidade"
-            id="city"
-            placeholder="Ex: São Paulo"
-            maxLength={50}
-            value={formData.city}
-            onChange={handleChange}
-            icon={<StateIcon />}
-            iconPosition="left"
-            disabled={!!zipCodeData}
-          />
-        </div>
+      <div className="flex flex-col gap-1 w-full text-left">
+        <label
+          htmlFor="tipoProjeto"
+          className="text-slate-700 font-semibold text-sm flex gap-1"
+        >
+          Tipo do projeto
+          <span className="text-red-500">*</span>
+        </label>
+        <select
+          id="tipoProjeto"
+          required
+          value={formData.tipoProjeto}
+          onChange={(e) => {
+            const nextType = e.target.value as WizardFormData["tipoProjeto"];
 
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          <Input
-            label="Estado"
-            id="state"
-            placeholder="Ex: São Paulo"
-            maxLength={50}
-            value={formData.state}
-            onChange={handleChange}
-            icon={<StateIcon />}
-            iconPosition="left"
-            disabled={!!zipCodeData}
-          />
-          <Input
-            label="UF"
-            id="stateCode"
-            placeholder="Ex: SP"
-            maxLength={2}
-            value={formData.stateCode}
-            onChange={handleChange}
-            icon={<StateIcon />}
-            iconPosition="left"
-            disabled={!!zipCodeData}
-          />
-        </div>
-
-        <Input
-          label="Telefone"
-          id="phone"
-          placeholder="(00) 00000-0000"
-          maxLength={30}
-          value={formData.phone}
-          onChange={handleChange}
-          icon={<PhoneIcon />}
-          iconPosition="left"
-        />
-
-        <InfoBox
-          title="Importante"
-          message="Essas informações serão visíveis no seu perfil público e utilizadas por empresas para entrar em contato com a sua ONG."
-        />
+            setFormData((prev) => ({
+              ...prev,
+              tipoProjeto: nextType,
+              metaCaptacao: nextType === "governamental" ? prev.metaCaptacao : "",
+            }));
+          }}
+          className={`w-full rounded-xl px-4 py-3 outline-none transition-all text-slate-900
+            border border-vinculo-gray bg-white
+            focus:border-vinculo-dark focus:ring-1 focus:ring-vinculo-dark
+            ${errors.tipoProjeto ? "!border !border-error focus:!border-error focus:!ring-error" : ""}`}
+        >
+          <option value="" disabled hidden>
+            Selecione
+          </option>
+          {PROJECT_TYPE_OPTIONS.map((option) => (
+            <option key={option.value} value={option.value}>
+              {option.label}
+            </option>
+          ))}
+        </select>
+        <p className="text-xs text-slate-500">
+          {formData.tipoProjeto
+            ? PROJECT_TYPE_OPTIONS.find(
+                (option) => option.value === formData.tipoProjeto,
+              )?.description
+            : "Escolha se o projeto terá ou não meta de captação."}
+        </p>
+        {errors.tipoProjeto && (
+          <p className="text-sm text-error" role="alert">
+            {errors.tipoProjeto}
+          </p>
+        )}
       </div>
+
+      {formData.tipoProjeto === "governamental" && (
+        <Input
+          id="metaCaptacao"
+          label="Meta de captação"
+          type="number"
+          min="0"
+          step="0.01"
+          placeholder="0,00"
+          value={formData.metaCaptacao}
+          onChange={(e) =>
+            setFormData((prev) => ({
+              ...prev,
+              metaCaptacao: e.target.value,
+            }))
+          }
+          error={errors.metaCaptacao}
+          isRequired
+        />
+      )}
+
+      <fieldset className="border-t border-slate-100 pt-5 border-x-0 border-b-0">
+        <legend className="text-vinculo-dark font-semibold text-base mb-1">
+          ODS{" "}
+          <span className="text-red-500" aria-hidden="true">
+            *
+          </span>
+        </legend>
+        <p className="text-sm text-slate-500 mb-4">
+          Selecione um ou mais ODS relacionados ao primeiro projeto.
+        </p>
+
+        <div className="grid grid-cols-1 gap-3">
+          {ODS_OPTIONS.map((option) => {
+            const selected = formData.odsProjeto.includes(option.value);
+
+            return (
+              <button
+                key={option.value}
+                type="button"
+                aria-pressed={selected}
+                onClick={() => toggleOds(option.value)}
+                className={`flex flex-col gap-1 rounded-xl border-2 px-4 py-3 text-left transition-all ${
+                  selected
+                    ? "border-vinculo-green bg-vinculo-green/10 text-vinculo-dark"
+                    : "border-vinculo-gray bg-white text-slate-600 hover:border-slate-300"
+                }`}
+              >
+                <span className="font-semibold text-sm">{option.label}</span>
+                <span className="text-xs text-slate-500">
+                  {option.description}
+                </span>
+              </button>
+            );
+          })}
+        </div>
+
+        {errors.odsProjeto && (
+          <p className="text-sm text-error mt-2" role="alert">
+            {errors.odsProjeto}
+          </p>
+        )}
+      </fieldset>
     </div>
   );
 }
