+++ frontend/src/types/wizard.types.ts
@@ -1,29 +1,16 @@
 export type OrganizationType = "npo" | "enterprise";
-
+export type ProjectType = "" | "social" | "governamental";
 export type WizardEsgOption = "ambiental" | "social" | "governanca";
 export type ProjectOdsOption = "1" | "2" | "3";
 
-/* Se precisar de campos opcionais utilize "pick". EXs:
-type SignupFields = Pick<
-  WizardFormData,
-  "nomeInstituicao" | "email" | "senha" | "confirmarSenha"
->;
-type NpoFields = Pick<WizardFormData, "cnpj">;
-type EnterpriseFields = Pick<WizardFormData, "razaoSocial">;
-*/
-
 export type WizardFormData = {
   nomeInstituicao: string;
-  email: string;
-  senha: string;
-  confirmarSenha: string;
   cnpj: string;
   razaoSocial: string;
   cpf: string;
   porteOng: "" | "pequena" | "media" | "grande";
   resumoInstitucional: string;
   esg: WizardEsgOption[];
-  // address
   zipCode: string;
   street: string;
   streetNumber: string;
@@ -33,6 +20,7 @@ export type WizardFormData = {
   stateCode: string;
   phone: string;
   nomeProjeto: string;
+  tipoProjeto: ProjectType;
   descricaoProjeto: string;
   metaCaptacao: string;
   odsProjeto: ProjectOdsOption[];
